LayoutTests/http/tests/security/contentSecurityPolicy/worker-blob-inherits-csp-expected.txt

CONSOLE MESSAGE: Refused to load the script 'http://localhost:8000/security/contentSecurityPolicy/resources/post-message.js' because it violates the following Content Security Policy directive: "script-src 'unsafe-inline' http://127.0.0.1:8000".


PASS Blob worker inherits CSP 
